AG-12173 Allow processing the .gitignore file (#58)
After the command npx @ag-grid-devtools/cli@32.0 migrate --from=31.2.1, error message Error: Untracked input files stop the process of migration. We can make it work by processing the .gitignore file

NOTE: this is a breaking change, files specified in .gitignore will NOT be processed anymore when running the command in a directory, agreed with Stephen that this is however the right behaviour - to decide if we want to alert the users or not in the release info

use "glob" and "ignore" to process correctly gitignore files recursively.
update e2e tests to verify gitignore files are ignored as expected