Publish 32.0.2 (#61)
* customizable allowed import libs, user configuration js file (#52)

- Adds the ability to define and load a js or ts configuration
- Expose a way in the configuration to register interceptor methods that can be used to allow custom modules and custom exports from custom libraries to be processed
- Fix the import.ts code to handle properly the configuration file interceptors
- Adds cli end to end tests

* fix build, add link to help screen (#57)

- Add link to the documentation when running migrate --help
- Add tsx as a dependency so typescript support for configuration is always available
- Fix the build configuration to properly emit user-config.cjs and user-config.d.ts
- Update some packages
- Fix user-config.ts so proper d.ts is generated correctly - the previous version was not generating a valid d.ts

* AG-12173 Allow processing the .gitignore file (#58)

After the command npx @ag-grid-devtools/cli@32.0 migrate --from=31.2.1, error message Error: Untracked input files stop the process of migration. We can make it work by processing the .gitignore file

NOTE: this is a breaking change, files specified in .gitignore will NOT be processed anymore when running the command in a directory, agreed with Stephen that this is however the right behaviour - to decide if we want to alert the users or not in the release info

use "glob" and "ignore" to process correctly gitignore files recursively.
update e2e tests to verify gitignore files are ignored as expected

* AG-11722 improve and fix commandline parsing (#59)

Fix command line parsing

The "--config=file" should not be included in the files to be processed
Partial semvers should be supported, "from=30" should work the same as "from=30.0.0"
All boolean flags should have an implicit "--no-flag" so they can be overridden when chaining commangs (normal cli behaviour)
--to= should support "latest" and partial semver, for example --to=32.1 or --to=latest

* release-32.0.2 (#60)

release-32.0.2

Updates in the migrate CLI:

- support for .gitignore files - when running in a folder without passing a list of files, .gitignore files will be honoured
- support for configuration file - this allows to define custom matchers to apply codemods also in wrapped or re-exported AG Grid components.